Show regnet and stockflow model diagrams
